+++ src/main/java/com/example/codeE/CodeEApplication.java
@@ -1,7 +1,10 @@
 package com.example.codeE;
 
+import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication
 public class CodeEApplication {
