+++ src/main/java/com/example/codeE/service/exercise/CodeExerciseService.java
@@ -10,6 +10,5 @@ public interface CodeExerciseService {
     void deleteCodeExerciseById(String exerciseId);
 
     CodeExercise updateCodeExercise(CodeExercise codeExercise);
-    
-    String runCodeExercise(String fileCodeContent, String language, String fileName);
+
 }
