+++ src/test/java/com/example/codeE/service/user/UserImplTest.java
@@ -2,7 +2,7 @@ package com.example.codeE.service.user;
 
 import com.example.codeE.model.user.User;
 import com.example.codeE.repository.UserRepository;
-import com.example.codeE.request.user.GetUsersRequest;
+import com.example.codeE.model.request.user.GetUsersRequest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
